+++ backend/src/main/java/ch/puzzle/okr/dto/goal/GoalDto.java
@@ -13,11 +13,12 @@ public class GoalDto {
     private String quarterLabel;
     private ExpectedEvolution expectedEvolution;
     private Unit unit;
-    private Long basicValue;
-    private Long targetValue;
+    private Double basicValue;
+    private Double targetValue;
 
     public GoalDto(GoalObjectiveDto objective, GoalKeyResultDto keyresult, Team team, Long progress,
-            String quarterLabel, ExpectedEvolution expectedEvolution, Unit unit, Long basicValue, Long targetValue) {
+            String quarterLabel, ExpectedEvolution expectedEvolution, Unit unit, Double basicValue,
+            Double targetValue) {
         this.objective = objective;
         this.keyresult = keyresult;
         this.teamId = team.getId();
@@ -94,19 +95,19 @@ public class GoalDto {
         this.unit = unit;
     }
 
-    public Long getBasicValue() {
+    public Double getBasicValue() {
         return basicValue;
     }
 
-    public void setBasicValue(Long basicValue) {
+    public void setBasicValue(Double basicValue) {
         this.basicValue = basicValue;
     }
 
-    public Long getTargetValue() {
+    public Double getTargetValue() {
         return targetValue;
     }
 
-    public void setTargetValue(Long targetValue) {
+    public void setTargetValue(Double targetValue) {
         this.targetValue = targetValue;
     }
 }
